syscall_user_dispatch.h: split out *_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

thread_info.h pulls in a lot of junk that sched.h that we don't need; in
particular, this helps to kill the printk.h dependency.

(cherry picked from commit 55b899aa3e7d0dc02ff9075b883d29eb2d0cb49a)
